Implemented DriverControlConfig
This encapsulates the hardware access to 8 input channels. These can be connected
to driver controlls or any kind of switches. 8 channels are managed by one
multiplexer. 4 Pins are needed, where one pin is the data pin and the other 3 pins
are the channel selectors. The bit selector setup for the channel mapping is stored
here as well.